+++ pkg/util/metadata/metadata.go
@@ -29,7 +29,7 @@ import (
 
 	"k8s.io/klog"
 
-	"k8s.io/cloud-provider-openstack/pkg/csi/cinder/mount"
+	"k8s.io/cloud-provider-openstack/pkg/util/mount"
 	"k8s.io/utils/exec"
 )
